Add Python bindings for `TokenizerOptions` and refactor Tokenizer implementation
- Introduced `TokenizerOptions` in Python bindings for enhanced configuration.
- Refactored `Tokenizer` implementation to leverage `TokenizerOptions`.
- Modularized Python bindings into `support`, `tokenizer_impl`, and `tokenizer_options`.
